// applications/panel-webapp/src/services/openapi/types.gen.ts
export type LowStockProduct = {
    /**
     * Product Code
     */
    product_code: string;
    /**
     * Product Name
     */
    product_name: string;
    /**
     * Quantity
     */
    quantity: number;
};

/**
 * ProductSchema
 */
export type ProductSchema = {
    /**
     * Product Code
     */
    product_code: string;
    /**
     * Product Name
     */
    product_name: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Price
     */
    price: number;
    /**
     * Quantity
     */
    quantity: number;
    /**
     * Category
     */
    category?: string | null;
    /**
     * Unit
     */
    unit?: string | null;
    /**
     * Shelf
     */
    shelf?: string | null;
    /**
     * Image Id
     */
    image_id?: Array<string>;
};

/**
 * ProductStatisticsResponse
 */
export type ProductStatisticsResponse = {
    /**
     * Total Products
     */
    total_products: number;
    /**
     * Total Quantity
     */
    total_quantity: number;
    /**
     * Total Categories
     */
    total_categories: number;
    /**
     * Low Stock Products
     */
    low_stock_products: Array<LowStockProduct>;
};